Supports VMs for premium storage, unmanaged disks
The added changes now allow an admin to select an image that is from a Premium Storage Account. Since the VM's disk is stored in the same storage account, there's no other issue, except adding logic to not log diagnostics wen using premium storage as diagnostics is not supported on premium.

Some extra work has been done to introduce managed disks, but that is not complete in this commit.