Proposed linked temlate changes
This is a larger change, splitting out the VM creation process into a linked template. Therefore, it includes passing all of the variables to it, changing the /dscextension to depend on the linked template, then adding language to call the right LinkedTemplate. In the LinkedTemplate itself, the logic was changed to just use the parameters, no variables.