Updated Script-RegisterSessionHost.ps1 to use ID of VM instead of name
Session hosts with a dns suffix different from the Active Directory domain name are not correctly found. Instead of passing $SessionHostName to Get-RdsSessionHost, we now get the ID of the VM from a local service and then get all session hosts in a pool and use the ID of the VM to match to the correct session host from the list of session hosts returned. Regardless of the computer name and dns suffix, the VM can be correctly matched by the ID.